updated SDK version. cleanup
