+++ package.json
@@ -16,7 +16,7 @@
     "format:check": "prettier './**/*.{js,ts,json}' --check"
   },
   "dependencies": {
-    "@connectifi/sdk": "0.1.1",
+    "@connectifi/sdk": "1.0.0",
     "@hubspot/api-client": "^12.0.1",
     "@middy/core": "^6.0.0",
     "@middy/http-header-normalizer": "^6.1.6",
