+++ workspaces/common/src/actions/data/hubspot/contacts.ts
@@ -1,16 +1,16 @@
 import {
     HubSpotSearchRequest,
     FilterOperatorEnum,
-    HSFilterGroup,
-    Query,
+    HubSpotFilterGroup,
     HubSpotContact,
-  } from '../../../types';
+  } from './common';
+  import { Query } from '../../../types';
   import { hubspotContactToFDC3 } from './common';
   import { Client } from "@hubspot/api-client";
   import { Context, Contact, ContactList } from '@finos/fdc3';
   
 
-  const appendToSearch  = ( filterGroup: Array<HSFilterGroup>, field: string, query: string) => {
+  const appendToSearch  = ( filterGroup: Array<HubSpotFilterGroup>, field: string, query: string) => {
 
     filterGroup.push({
         filters:[
@@ -32,6 +32,16 @@ import {
     });
 };
 
+const isCompanyContext = (context: Context): boolean => {
+    if (context.type  === 'fdc3.instrument' && context.name){
+        return true;
+    }
+    if (context.type  === 'connect.company' && context.name){
+        return true;
+    }
+    return false;
+};
+
 export const searchContacts = async (apiKey: string, context: Context): Promise<ContactList> => {
     const hubspotClient = new Client({ accessToken: apiKey });
     //get the contacts
@@ -39,7 +49,7 @@ export const searchContacts = async (apiKey: string, context: Context): Promise<
 
     }
 
-    const filterGroups: Array<HSFilterGroup>  = [];
+    const filterGroups: Array<HubSpotFilterGroup>  = [];
     
     if (context.type === 'fdc3.contact'){
         const contact = context as Contact;
@@ -94,20 +104,22 @@ export const searchContacts = async (apiKey: string, context: Context): Promise<
 
     }
     
-    if (context.type  === 'fdc3.instrument' && context.name){
-        //find the company 
+    if (isCompanyContext(context)){
+        //find the company
+        //match on name in a 'fuzzy' way
+        const nameParts = context.name.split(' ');
+        const  filters = [];
+        nameParts.forEach((part) => {
+            filters.push({
+            propertyName:'name',
+            operator: FilterOperatorEnum.ContainsToken,
+            value: `${nameParts[0]}*`
+            });
+        });
+        
         const companies = await hubspotClient.crm.companies.searchApi.doSearch({
             filterGroups:[
-                {
-                    filters: [
-                        {
-                            propertyName:'name',
-                            operator: FilterOperatorEnum.ContainsToken,
-                            value: `${context.name}*`
-                        }
-                    ]
-                    
-                }
+                { filters }
             ]
         });
         if (companies.results[0]){
@@ -129,8 +141,6 @@ export const searchContacts = async (apiKey: string, context: Context): Promise<
             filterGroups,
             limit: 100
         });
-
-
 
         return {
                 type: 'fdc3.contactList',
